Send a heartbeat on establishing a stream
Summary:
On establishing a stream, we send a heartbeat. This paves the way for
the fully fledged heartbeat feature. It also solves the problem when
reconnecting to a machine and the stream is silent, we cannot determine
the connection is healthy and propagate the information to the observer.

Depends on D60171.

Test Plan: make -j all check

Reviewers: pja, #rocketspeed

Reviewed By: pja, #rocketspeed

Subscribers: pja

Differential Revision: https://reviews.facebook.net/D60381