Opaque user data per subscription
Summary:
Introduces a user_data pointer to subscriptions. The high-level goal is to remove the reliance on defined object layout in the subscriptions map so that we can use alternative data structures and storage systems (e.g. RocksDB).

This fixes things for proxy2. Next diff will fix up the client, then I'll be able to remove the template and introduce the fully abstracted interface.

Test Plan: make -j check all

Reviewers: #rocketspeed, stupaq-fb

Reviewed By: #rocketspeed, stupaq-fb

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D60939