Disable java tests
Summary: TestPilot cannot find rsjni, and I can't figure out how to make it build it and find it. If someone knows how then please shout, but it doesn't seem worth spending time on at this stage since no one uses the java client. If that changes, we can invest time in a fix.

Test Plan: n/a

Reviewers: #rocketspeed, rishijhelumi

Reviewed By: #rocketspeed, rishijhelumi

Subscribers: stupaq-fb, dhruba

Differential Revision: https://reviews.facebook.net/D63753