Fix connection failure assertion
Summary: If we fail to open a new stream, the resilient receiver asserts. It seems that the failure path for reconnect() inside ResilientStreamReceiver::Reconnect doesn't return and tries to fall through to setup a retry. However, reconnect() already sets up the retry on failure, so it doesn't seem necessary, and I think we can just return whether it succeeds or not.

Test Plan:
```$ ./centrifuge_client --mode=subscribe-unsubscribe-rapid --num_subscriptions=999999999
$ ./accept_all_rocketeer
# fails after ~2^15 reconnects```

Reviewers: gregsexton

Reviewed By: gregsexton

Subscribers: stupaq-fb, dhruba, #rocketspeed

Differential Revision: https://reviews.facebook.net/D64179