Move heartbeat timeout to socket
Summary:
We monitor streams for heartbeats at the socket. We also timeout here
and notify a stream if it is unhealthy. This relies on the assumption
that a socket cannot fail independently from stream notification. This
seems fair and should be a pragmatic perf win.

Test Plan: make -j all check

Reviewers: pja, #rocketspeed

Reviewed By: pja, #rocketspeed

Subscribers: pja

Differential Revision: https://reviews.facebook.net/D62991