Implement client-side ack logic (#149)
* Implement client-side ack logic

* Only drop updates from the same source

* Fixup nits

* Zero-out sub id in MessageSubAck

* Make sub acks a first class state in the machine

Rather than a piece of metadata, reduce memory overhead by making
'waiting on ack' a state. We transition:

pending sub -> pending ack -> synced

The consequence is that we do not accept deliveries for subs
pending-ack. This is slightly more aggressive and could have edge cases
yet discovered. It has meant a more invasive change.

* Fix comments

* Only accept unsub when fully synced

For safety, we only accept unsubs from the server when a subscription
has been fully synced (hence ack'd). This changes the server contract to
require to ack before any deliveries or terminations.